\name{CYBERTRACK2-package}
\alias{CYBERTRACK2-package}
\alias{CYBERTRACK2}
\docType{package}
\title{
\packageTitle{CYBERTRACK2}
}
\description{
\packageDescription{CYBERTRACK2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CYBERTRACK2}
\packageIndices{CYBERTRACK2}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CYBERTRACK2}

Maintainer: \packageMaintainer{CYBERTRACK2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
